image-uploader/requirements.txt
djangorestframework==3.7.7
Pillow==5.0.0
virtualenv==15.1.0
